requirements.txt
aiocache==0.11.1
aiohttp==3.8.1
aioredis==1.3.1
aiosignal==1.2.0
alembic==1.8.0
anyio==3.6.1
apispec==5.2.2
apispec-webframeworks==0.5.2
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
Babel==2.10.2
backoff==2.1.2
blinker==1.4
certifi==2022.6.15
cfgv==3.3.1
charset-normalizer==2.1.0
click==8.1.3
clickhouse-driver==0.2.4
Deprecated==1.2.13
distlib==0.3.6
dnspython==2.2.1
elasticsearch==7.9.1
email-validator==1.2.1
fastapi==0.61.1
filelock==3.8.0
flake8==4.0.0
flasgger==0.9.5
Flask==2.1.2
Flask-BabelEx==0.9.4
Flask-JWT-Extended==4.4.1
Flask-Login==0.6.1
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
flask-swagger-ui==4.11.1
Flask-WTF==1.0.1
frozenlist==1.3.1
gevent==21.12.0
googleapis-common-protos==1.56.4
greenlet==1.1.2
grpcio==1.47.0
gunicorn==20.1.0
h11
hiredis==2.0.0
httpcore==0.15.0
httptools==0.4.0
httpx==0.23.0
identify==2.5.3
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.6.0
kafka-python==2.0.2
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.16.0
mccabe==0.6.1
mistune==2.0.2
multidict==6.0.2
mypy==0.950
mypy-extensions==0.4.3
nodeenv==1.7.0
opentelemetry-api==1.10.0
opentelemetry-exporter-jaeger==1.10.0
opentelemetry-exporter-jaeger-proto-grpc==1.10.0
opentelemetry-exporter-jaeger-thrift==1.10.0
opentelemetry-instrumentation==0.29b0
opentelemetry-instrumentation-flask==0.29b1
opentelemetry-instrumentation-wsgi==0.29b0
opentelemetry-sdk==1.10.0
opentelemetry-semantic-conventions==0.29b0
opentelemetry-util-http==0.29b0
orjson==3.8.0
outcome==1.2.0
packaging==21.3
passlib==1.7.4
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.19.0
protobuf==4.21.2
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pydantic==1.9.0
pyflakes==2.4.0
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==6.2.5
pytest-asyncio==0.12.0
python-dotenv==0.20.0
python-json-logger==2.0.4
python-logstash==0.4.8
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
redis==4.3.3
requests==2.28.1
rfc3986==1.5.0
sentry-sdk==1.9.5
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
speaklater==1.3
SQLAlchemy==1.4.37
starlette==0.13.6
thrift==0.16.0
toml==0.10.2
tomli==2.0.1
trio==0.21.0
trio-asyncio==0.12.0
types-redis==4.2.0
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.11
uvicorn==0.17.6
uvloop==0.16.0
virtualenv==20.16.3
Werkzeug==2.1.2
wrapt==1.14.1
WTForms==3.0.1
yarl==1.8.1
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
